+++ kodando-react-dom/src/main/kotlin/kodando/react/dom/extensions.kt
@@ -3,7 +3,8 @@ package kodando.react.dom
 import kodando.react.ElementBuilder
 import kodando.react.PropSetter
 import kodando.react.ReactElement
-import kodando.react.objectOf
+import kodando.react.objectWithShapeOf
+import kodando.react.objectWithShapeBy
 
 /**
  * Created by danfma on 19/01/17.
@@ -21,5 +22,5 @@ operator fun <TProps : HTMLElementAttributes> ElementBuilder<TProps>.invoke(clas
 	}
 
 fun styleOf(setter: CSSProperties.() -> Unit): CSSProperties {
-	return objectOf(setter)
+	return objectWithShapeBy(setter)
 }