+++ kodando-react-router/src/main/kotlin/kodando/react/router/dom/Prompt.kt
@@ -1,4 +1,4 @@
-package kodando.react.router
+package kodando.react.router.dom
 
 import kodando.react.*
 
@@ -8,7 +8,7 @@ import kodando.react.*
 
 @JsName("promptWithText")
 fun prompt(message: String, matched: Boolean = false): ReactElement? {
-    return React.createElement(Module.PromptClass, propsBy<PromptProps> {
+    return React.createElement(Module.PromptClass, unsafePropsBy<PromptProps> {
         this.messageText = message
         this.matched = matched
     })
@@ -17,8 +17,8 @@ fun prompt(message: String, matched: Boolean = false): ReactElement? {
 
 @JsName("promptWithFunc")
 fun prompt(messageFactory: (Location) -> String, matched: Boolean = false): ReactElement? {
-    return React.createElement(Module.PromptClass, propsBy<PromptProps> {
-        this.messageFunc = messageFunc
+    return React.createElement(Module.PromptClass, unsafePropsBy<PromptProps> {
+        this.messageFunc = messageFactory
         this.matched = matched
     })
 }