+++ kodando-react-router/src/main/kotlin/kodando/react/router/dom/Redirect.kt
@@ -1,4 +1,4 @@
-package kodando.react.router
+package kodando.react.router.dom
 
 import kodando.react.*
 
@@ -9,7 +9,7 @@ import kodando.react.*
 
 @JsName("redirectToUrl")
 fun redirect(to: String, from: String? = null, push: Boolean? = null): ReactElement? {
-    return React.createElement(Module.RedirectClass, propsBy<RedirectProps> {
+    return React.createElement(Module.RedirectClass, unsafePropsBy<RedirectProps> {
         this.toUrl = to
         this.from = from ?: undefined
         this.push = push ?: undefined
@@ -19,7 +19,7 @@ fun redirect(to: String, from: String? = null, push: Boolean? = null): ReactElem
 
 @JsName("redirectToLocation")
 fun redirect(location: Location, from: String? = null, push: Boolean? = null): ReactElement? {
-    return React.createElement(Module.RedirectClass, propsBy<RedirectProps> {
+    return React.createElement(Module.RedirectClass, unsafePropsBy<RedirectProps> {
         this.toLocation = location
         this.from = from ?: undefined
         this.push = push ?: undefined